+++ api/spec/controllers/spree/api/checkouts_controller_spec.rb
@@ -63,10 +63,26 @@ module Spree
         json_response['state'].should == 'cart'
       end
 
-      it "should transition a recently created order from cart do address" do
+      it "should transition a recently created order from cart to address" do
         order.state.should eq "cart"
         order.email.should_not be_nil
         api_put :update, :id => order.to_param
+        order.reload.state.should eq "address"
+      end
+
+      it "can take line_items_attributes as a parameter" do
+        line_item = order.line_items.first
+        api_put :update, :id => order.to_param, :order_token => order.token,
+                         :order => { :line_items_attributes => { line_item.id => { :quantity => 1 } } }
+        response.status.should == 200
+        order.reload.state.should eq "address"
+      end
+
+      it "can take line_items as a parameter" do
+        line_item = order.line_items.first
+        api_put :update, :id => order.to_param, :order_token => order.token,
+                         :order => { :line_items => { line_item.id => { :quantity => 1 } } }
+        response.status.should == 200
         order.reload.state.should eq "address"
       end