+++ api/CHANGELOG.md
@@ -39,3 +39,7 @@
 * The Shipments API no longer returns inventory units. Instead, it will return manifest objects. This is necessary due to the split shipments changes brought in by Spree 2.
 
     * Ryan Bigg
+
+* Checkouts API's update action will now correctly process line item attributes (either `line_items` or `line_item_attributes`)
+
+    * Ryan Bigg