+++ api/CHANGELOG.md
@@ -1,5 +1,9 @@
-## Spree 2.0.3 (unreleased) ##
+## Spree 2.0.4 (unreleased)
 
 * PUT requests to Checkouts API endpoints now require authorization to alter an order.
 
     *Ryan Bigg*
+
+* Checkouts API's update action will now correctly process line item attributes (either `line_items` or `line_item_attributes`)
+
+    * Ryan Bigg*